move files of repair
