+++ satellite/repair/repair.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package datarepair
+package repair
 
 // Empty file to workaround coverpkg issue https://golang.org/issue/27333
